Abstract MessagingController into a more generic class
