+++ hmi.py
@@ -1,10 +1,7 @@
 #!/bin/env python3
 
-import socket
 import commsettings
-import time
 import messaging
-import queue
 import json
 import random
 import logging
@@ -29,49 +26,6 @@ if not IMPORT_UI_ONTHEFLY:
 else:
     class Ui_MainWindow:
         pass
-
-
-class HMIMessageController(QThread):
-    signal_recieve_message = pyqtSignal(str)
-
-    def __init__(self, loglevel=logging.INFO):
-        QThread.__init__(self)
-        self.logger = logs.build_logger(__name__, loglevel)
-        self.loglevel = loglevel
-
-        self.receiver = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.clientqueue = queue.Queue()
-        self.msg_controller = messaging.Messaging(commsettings.MESSAGE_BREAKER, self.receiver, self.clientqueue,
-                                                                                loglevel=self.loglevel, name="HMILogic")
-
-    def run(self):
-        # Configure Socket to allow reuse of sessions in TIME_WAIT. Otherwise, "Address already in use" is encountered
-        # Per suggestion on https://stackoverflow.com/questions/29217502/socket-error-address-already-in-use/29217540
-        # by ForceBru
-        self.receiver.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        self.receiver.bind(("127.0.0.1", commsettings.HMI_LISTEN))
-        self.receiver.listen(2)
-        self.logger.info("successfully opened port " + str(commsettings.HMI_LISTEN))
-        # Keep trying to create the sender until the correct receiver has been created
-        while True:
-            try:
-                self.sender = socket.create_connection(("127.0.0.1", commsettings.GAME_HMI_LISTEN))
-                break
-            except Exception as e:
-                self.logger.error(e)
-                time.sleep(1)
-                continue
-
-        self.msg_controller.start()
-
-        while True:
-            if not self.clientqueue.empty():
-                self.signal_recieve_message.emit(self.clientqueue.get())
-            time.sleep(.1)
-
-    @pyqtSlot(str)
-    def send_message(self, msg):
-        self.msg_controller.send_string(self.sender, msg)
 
 
 class HMILogicController(QObject):
@@ -323,7 +277,10 @@ class HMI(QtWidgets.QMainWindow, Ui_MainWindow):
             "Double" : QSound("Double.wav")
         }
 
-        self.MSG_controller = HMIMessageController(loglevel=loglevel)
+        self.MSG_controller = messaging.MessageController(loglevel=loglevel,
+                                                          msg_controller_name="HMILogic",
+                                                          listen_port=commsettings.HMI_LISTEN,
+                                                          target_port=commsettings.GAME_HMI_LISTEN)
 
         self.registration_wizard = wizard.MyWizard(ui_file="register_user_wizard.ui", loglevel=self.loglevel)
